Please read below
I re-read the assignment and I have to admit that I missed the very last part
of it when I had first read it (no excuses, just saying).  The very last part
states that you'd like the parser reusable. My bad - fixed.

Few points: I inspired myself with the `database/sql` Row's Next()/Scan()
methods and modeled the parser after it.  Like `database/sql` I don't force new
allocations with Scan onto the user of this API.  Tests are included.

As for the DDD for port-domain I have to admit that it took me a while to
realize:

First I wrote additional interface for the bulk handler and when I had been
looking at it for a while, it hit me like a ton of bricks. It looked exactly
like my previous PortStorer :facepalm: So I deleted both of them and moved my
PortStorer into `ports` package and removed the `Port` part (it's implicit from
the package name) Then I implemented a small wrapper around the
generated grpc client for client-api to be able to send to port-domain.
Furthermore I also isolated code into own packages in both services (handlers,
service, grpc...)

From the ddd/code structure PoV i think this is what you wanted to see...

However, there is one slight small detail why I didn't put the same interface
into client-api on my first try.  When the assignment stated that the
client-api has only 200 MB and the file has to be read item-by-item (otherwise
it wouldn't fit into memory), then listing and encoding should have
been done one-by-one from  grpc stream as well, because it too might not fit
into memory. The cardinality is the same.  But I guess that if the
structure/ddd was the goal of the homework then this doesn't really matter that
much.  In a real world scenario we would either a) have a different interface for
client-api or b) have the same interface, but the listing would be an iterator-like
(like the parser api).